Move base class to different folders from auto-generated files
